CALL-87 (#7)
* CALL-87 WIP

* CALL-87 WIP

* CALL-87 WIP

* CALL-87 WIP

* CALL-87 WIP

* CALL-87 WIP

* CALL-87

---------

+++ backend/app/api/routes/users.py
@@ -1,56 +1,23 @@
-import uuid
-from typing import Any
+from fastapi import APIRouter
 
-from fastapi import APIRouter, Depends, HTTPException, status
-from sqlmodel import col, delete, func, select
+from app.errors import (
+    RequestConflictError,
+    BadRequestError,
+    AccessDeniedError
+)
 
 from app import crud
 
-from app.api.deps import (
-    CurrentUser,
-    SessionDep,
-    get_current_active_superuser,
-)
+from app.api.deps import CurrentUser, SessionDep
 
 from app.core.security import get_password_hash, verify_password
 
-
 from app.models.message import Message
-from app.models.password import UpdatePassword
-from app.models.user import (
-    User,
-    UserCreate,
-    UserPublic,
-    UserRegister,
-    UsersPublic,
-    UserUpdate,
-    UserUpdateMe,
-)
+from app.models.password import UpdatePassword, CheckPassword
+from app.models.user import UserPublic, UserUpdateMe
 
 
 router = APIRouter(prefix="/users", tags=["users"])
-
-
-# @router.get(
-#     "/",
-#     dependencies=[Depends(get_current_active_superuser)],
-#     response_model=UsersPublic,
-# )
-# def read_users(session: SessionDep, skip: int = 0, limit: int = 100) -> Any:
-#     """
-#     Retrieve users.
-#     """
-
-#     count_statement = select(func.count()).select_from(User)
-#     count = session.exec(count_statement).one()
-
-#     statement = (
-#         select(User).order_by(
-#             col(User.created_at).desc()).offset(skip).limit(limit)
-#     )
-#     users = session.exec(statement).all()
-
-#     return UsersPublic(data=users, count=count)
 
 
 @router.patch("/me", response_model=UserPublic)
@@ -62,15 +29,12 @@ def update_current_user(
     """
 
     if user_in.email:
-        existing_user = crud.get_user_by_email(
-            session=session, email=user_in.email
+        existing_user = crud.get_user_by_username(
+            session=session, username=user_in.email
         )
 
         if existing_user and existing_user.id != current_user.id:
-            raise HTTPException(
-                status_code=status.HTTP_409_CONFLICT,
-                detail="User with this email already exists"
-            )
+            raise RequestConflictError("User with this email already exists")
 
     user_data = user_in.model_dump(exclude_unset=True)
     current_user.sqlmodel_update(user_data)
@@ -93,14 +57,11 @@ def update_password(
     )
 
     if not verified:
-        raise HTTPException(
-            status_code=status.HTTP_400_BAD_REQUEST, detail="Incorrect password"
-        )
+        raise BadRequestError("Incorrect password")
 
     if body.current_password == body.new_password:
-        raise HTTPException(
-            status_code=status.HTTP_400_BAD_REQUEST,
-            detail="New password cannot be the same as the current one"
+        raise BadRequestError(
+            "New password cannot be the same as the current one"
         )
 
     hashed_password = get_password_hash(body.new_password)
@@ -120,92 +81,23 @@ def get_current_user(current_user: CurrentUser) -> UserPublic:
 
 
 @router.delete("/me", response_model=Message)
-def delete_current_user(session: SessionDep, current_user: CurrentUser) -> Message:
+def delete_current_user(session: SessionDep, body: CheckPassword, current_user: CurrentUser) -> Message:
     """
     Delete own user.
     """
     if current_user.is_superuser:
-        raise HTTPException(
-            status_code=status.HTTP_403_FORBIDDEN,
-            detail="Super users are not allowed to delete themselves"
+        raise AccessDeniedError(
+            "Super users are not allowed to delete themselves"
         )
+
+    verified, _ = verify_password(
+        body.password, current_user.hashed_password
+    )
+
+    if not verified:
+        raise BadRequestError("Incorrect password")
 
     session.delete(current_user)
     session.commit()
 
     return Message(message="User deleted successfully")
-
-
-# @router.get("/{user_id}", response_model=UserPublic)
-# def read_user_by_id(
-#     user_id: uuid.UUID, session: SessionDep, current_user: CurrentUser
-# ) -> Any:
-#     """
-#     Get a specific user by id.
-#     """
-#     user = session.get(User, user_id)
-#     if user == current_user:
-#         return user
-#     if not current_user.is_superuser:
-#         raise HTTPException(
-#             status_code=403,
-#             detail="The user doesn't have enough privileges",
-#         )
-#     if user is None:
-#         raise HTTPException(status_code=404, detail="User not found")
-#     return user
-
-
-# @router.patch(
-#     "/{user_id}",
-#     dependencies=[Depends(get_current_active_superuser)],
-#     response_model=UserPublic,
-# )
-# def update_user(
-#     *,
-#     session: SessionDep,
-#     user_id: uuid.UUID,
-#     user_in: UserUpdate,
-# ) -> Any:
-#     """
-#     Update a user.
-#     """
-
-#     db_user = session.get(User, user_id)
-#     if not db_user:
-#         raise HTTPException(
-#             status_code=404,
-#             detail="The user with this id does not exist in the system",
-#         )
-#     if user_in.email:
-#         existing_user = crud.get_user_by_email(
-#             session=session, email=user_in.email)
-#         if existing_user and existing_user.id != user_id:
-#             raise HTTPException(
-#                 status_code=409, detail="User with this email already exists"
-#             )
-
-#     db_user = crud.update_user(
-#         session=session, db_user=db_user, user_in=user_in)
-#     return db_user
-
-
-# @router.delete("/{user_id}", dependencies=[Depends(get_current_active_superuser)])
-# def delete_user(
-#     session: SessionDep, current_user: CurrentUser, user_id: uuid.UUID
-# ) -> Message:
-#     """
-#     Delete a user.
-#     """
-#     user = session.get(User, user_id)
-#     if not user:
-#         raise HTTPException(status_code=404, detail="User not found")
-#     if user == current_user:
-#         raise HTTPException(
-#             status_code=403, detail="Super users are not allowed to delete themselves"
-#         )
-#     statement = delete(Item).where(col(Item.owner_id) == user_id)
-#     session.exec(statement)
-#     session.delete(user)
-#     session.commit()
-#     return Message(message="User deleted successfully")
